Release v0.8.4
Release v0.8.4

(cherry picked from commit 9ee4897dd901d7c82d8fb4f43bfcb94e36fd9be3)

+++ drivers/input/sensors/smi230/smi230.h
@@ -683,7 +683,7 @@ int8_t smi230_gyro_read_fifo_data(struct smi230_fifo_frame *fifo,
  * @retval SMI230_E_NULL_PTR - Error: Null pointer error
  */
 int8_t smi230_gyro_extract_fifo(struct smi230_sensor_data *gyro_data,
-				uint8_t *fifo_length,
+				uint16_t *fifo_length,
 				struct smi230_fifo_frame *fifo,
 				const struct smi230_dev *dev);
